extract the fetching logic from SubsectionsMenu
create a wrapper around SubsectionMenu to handle the fetching logic
