+++ src/features/Nav/Nav.js
@@ -3,7 +3,7 @@ import styled from 'styled-components/macro';
 import { Route } from 'react-router-dom';
 
 import SectionMenu from 'features/SectionMenu/SectionMenu';
-import SubsectionMenu from 'features/SubsectionMenu/SubsectionMenu';
+import SubsectionsContainer from 'features/SubsectionMenu/SubsectionsContainer';
 
 const StyledNav = styled.nav`
     box-shadow: 0 0 8px 1px var(--black);
@@ -17,7 +17,7 @@ const Nav = () => (
     <StyledNav>
         <SectionMenu />
         <Route path='/:sectionName?/:subsecName?/:featureName?'>
-            <SubsectionMenu />
+            <SubsectionsContainer />
         </Route>
     </StyledNav>
 )
